backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic==2.10.0
python-multipart==0.0.9
pymysql
sqlalchemy
docker
pydantic[email]
